requirements.txt
Django==1.10
django-cors-middleware==1.3.1
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
wheel==0.24.0
